backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-groq
langgraph
python-dotenv
pydantic
pydantic-settings
